# Conflicts:
#	alpha/src/main/java/com/icourt/alpha/fragment/TabNewsFragment.java
#	alpha/src/main/java/com/icourt/alpha/service/SyncDataService.java

+++ alpha/src/main/java/com/icourt/alpha/fragment/ContactListFragment.java
@@ -291,7 +291,7 @@ public class ContactListFragment extends BaseFragment implements BaseRecyclerAda
      * @param data
      */
     private void insertAsynContact(List<GroupContactBean> data) {
-        if (data == null) return;
+        if (data == null||data.isEmpty()) return;
         try {
             contactDbService.deleteAll();
             contactDbService.insertOrUpdateAsyn(new ArrayList<IConvertModel<ContactDbModel>>(data));
